+++ lib/constants/helpers.dart
@@ -1,5 +1,0 @@
-String enumToString<T>(T e) => e.toString().split('.').last;
-
-T enumFromString<T>(String str, List<T> values) => values.firstWhere(
-      (e) => str == enumToString(e),
-    );