+++ lib/extensions/enum_extension.dart
@@ -1,0 +1,1 @@
+String enumToString<T>(T e) => e.toString().split('.').last;