+++ lib/services/firestore_database.dart
@@ -1,24 +1,14 @@
 import 'package:cccc/models/plaid/account.dart';
 import 'package:cccc/models/plaid/transaction.dart';
 import 'package:cccc/models/user.dart';
-import 'package:cccc/services/firebase_auth.dart';
 import 'package:cccc/services/firestore_path.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'firestore_service.dart';
 
-final databaseProvider = Provider.family<FirestoreDatabase?, String?>(
-  (ref, id) {
-    final auth = ref.watch(authProvider);
-    final uid = auth.currentUser?.uid;
-
-    if (id != null) {
-      return FirestoreDatabase(uid: id);
-    } else if (uid != null) {
-      return FirestoreDatabase(uid: uid);
-    } else {
-      return null;
-    }
+final databaseProvider = Provider.family<FirestoreDatabase, String>(
+  (ref, uid) {
+    return FirestoreDatabase(uid: uid);
   },
 );
 
@@ -62,14 +52,15 @@ class FirestoreDatabase {
         toBuilder: (model) => model.toMap(),
       );
 
-  Stream<List<Transaction?>> transactionsStream() =>
-      _service.collectionStream<Transaction>(
-        path: FirestorePath.transactions(uid!),
-        fromBuilder: (json, id) => Transaction.fromMap(json!),
-        toBuilder: (model) => model.toMap(),
-        orderByField: 'date',
-        descending: true,
-      );
+  Stream<List<Transaction?>> transactionsStream() {
+    return _service.collectionStream<Transaction>(
+      path: FirestorePath.transactions(uid!),
+      fromBuilder: (json, id) => Transaction.fromMap(json!),
+      toBuilder: (model) => model.toMap(),
+      orderByField: 'date',
+      descending: true,
+    );
+  }
 
   Stream<List<Account?>> accountsStream() => _service.collectionStream<Account>(
         path: FirestorePath.accounts(uid!),
@@ -78,4 +69,12 @@ class FirestoreDatabase {
         orderByField: 'type',
         descending: false,
       );
+
+  Stream<Account?> accountStream(String accountId) {
+    return _service.documentStream<Account>(
+      path: FirestorePath.account(uid!, accountId),
+      fromBuilder: (json, id) => Account.fromMap(json!),
+      toBuilder: (model) => model.toMap(),
+    );
+  }
 }